requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
django-crispy-forms==2.0
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
